refactor: orders wave 6 slice B — extract order_suggestion_probe (#4246)
Move incremental validator construction, probe template, test hooks, and
is*OrderAccepted wrappers into order_suggestion_probe.dart; slim
order_suggestion_context.dart to logger + diplomatic trial helper with
re-exports. Clear context from shrink-only grandfather.

Refs #4246
